refactor: 패키지 구조 수정(#31)
- promotion 따로 분리
