+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/req/promotion/ReqPostPromotionDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.product.application.dtos.req.product;
+package com.sparta.hotdeal.product.application.dtos.req.promotion;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotNull;
