+++ product/src/main/java/com/sparta/hotdeal/product/infrastructure/repository/promotion/PromotionRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.sparta.hotdeal.product.infrastructure.repository.promotion;
 
-import com.sparta.hotdeal.product.domain.entity.product.Promotion;
-import com.sparta.hotdeal.product.domain.entity.product.PromotionStatusEnum;
+import com.sparta.hotdeal.product.domain.entity.promotion.Promotion;
+import com.sparta.hotdeal.product.domain.entity.promotion.PromotionStatusEnum;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
